requirements.txt
pyttsx3
google-generativeai
speech_recognition  \azaa\